Fixed broken batch size math.
When separating the dataset into batches, we round off to
make every batch the same size. The MPI logic further splits
each batch amoung the ranks, which means recalculating the
size of each batch. I thought one expression would work in both
cases, but it doesn't due to the roundoff error.

This breaks the MPI code, I'll fix it later.